Bring in features and fixes from next
* Refactor media service and serialize GSMTC operations (#32)

Refactors the media service to serialize operations and prevent a thread stampede when GSMTC is blocked for the process.

- Introduces optimistic state presentation for faster UI updates.
- Hardens utility classes to avoid lifecycle, timing and synchronization issues and ensure proper disposal.

* Improve volume commands and underlaying infrastructure (#33)

Re-imagine volume support and turn it native

- Improves performance of volume commands.
- Adds new top-level volume commands.
- Adds a new compact volume control to Media controls page and allows it to be pinned to the dock.
- Replaces inner implementation with native AOT safe one.
- Enables AOT publish and trimming.
- Adds a global settings toggle to disable volume management entirely.

* Improve layout and formatting of the settings page (#35)

Improve layout and formatting of the settings page, add sections and formatted text block

* Add media details to now playing and media list items and add full metadata page (#36)

* Add icon themes (#37)

- Introduces the icon service that dynamically provides icons based on user settings.
- Adds several new icon themes (Fluent outline, Fluent solid, Pink princess, Chubby, Neon).

* Fix issues caused by refactor

- Adds per-source playback action queue: presses record the desired end state, a single worker coalesces spam to the newest action (generation-guarded).
- Returns optimistic toast immediately from play/pause press; press path no longer waits on the GSMTC gate.
- Resolves play/pause from the optimistic display state instead of the lagging GSMTC snapshot; always send the command (drop false-success early-outs).
- Tracks CanPlay and treat "playing with Play still enabled" as a transitional snapshot resolving to Pause; fixes "Play → Stop → Pause" title flash.
- Adds GsmtcErrors classifier for E_BOUNDS / RO_E_CLOSED (WinRT projections of dead-session failures).
- Retry a queued playback action once after a stale session: raise SessionInvalidated, let the service refresh rebind the session.
- Harden RefreshCore: skip dead sessions instead of aborting the refresh, pass AUMID into MediaSource (no native re-reads), retry failed refresh up to 3× with 500 ms delay
Isolate PauseOthersOnPlay per session; one dead session no longer aborts play or session switch.
Fail one-shot session commands with a toast on unexpected errors instead of leaking the exception
- Uses cached AUMID in MediaSource logging and app-info lookup; the native re-read inside catch blocks itself threw E_BOUNDS on dead sessions (the bare exception seen in the log).

* Fix the fix

* Refactor class hierarchy a bit

* Minor refactor

* Use Fluent icons for the volume control context menu and as a default fallback

* Update readme and docs

* Separate package project (WAP) and upgrade to .NET 10

* Improve realiability of metadata updates

* Fix initial dock band labels (skip track items regressed and displayed title by default)

* Allow use to select behavior of first dock item (bring app to front, open main page, open metadata info)

* Add separators to context menus

* Add Ctrl+I keyboard shortcut to metadata info commands

* Update localization

* Improve publishing project

* Switch the logging to Microsoft.Extension.Logging and a newer JPSoftworks.CommandPalette.Extensions.Toolkit

* Separate Volume zero and Volume unmute icons

* Enable media thumbnails by default

* Cleanup Icons class

* Convert prev/next track fallback commands to normal commands

* Add System volume dock band to the flyout

* Update localization

* Add a 'Report a problem' page

* Improve publishing project

* Update change log and readme

+++ src/MediaControlsExtension/Interop/PROPERTYKEY.cs
@@ -15,15 +15,15 @@ namespace JPSoftworks.MediaControlsExtension.Interop;
 public struct PROPERTYKEY : IEquatable<PROPERTYKEY>
 {
     public Guid fmtid;
-    public nuint pid;
+    public uint pid;
 
-    private PROPERTYKEY(Guid fmtid, nuint pid)
+    private PROPERTYKEY(Guid fmtid, uint pid)
     {
         this.fmtid = fmtid;
         this.pid = pid;
     }
 
-    public static PROPERTYKEY FromString(string fmtid, nuint pid)
+    public static PROPERTYKEY FromString(string fmtid, uint pid)
     {
         return new PROPERTYKEY(Guid.Parse(fmtid), pid);
     }
